Allow multiple input styles for node coordinates
NumPy to the rescue! Multiple input styles are now supported for node
coordinates. The signature takes a variable number of arguments, but
they will all be flattened into a 1-D NumPy array. So all of the
following examples work::

    >>> Node(0, 0.0, 0.0).tcl_code()
    'node 0 0 0'
    >>> Node(1, 1.0).tcl_code()
    'node 1 1'
    >>> Node(2, [0.0, 1.0, 2.0], mass=[0]*6).tcl_code()
    'node 2 0 1 2 -mass 0 0 0 0 0 0'

This is done simply by converting the given inputs -- whatever they are!
-- to a NumPy array and calling `flatten`. `mass` is only converted if
not None.

A dtype could also be specified... hmm... coerce to floats, yes...