Add tclescape function and improve tcllist
tclescape() will escape meta-characters so that the given string will be
parsed as a plain string by Tcl.

tcllist() now uses tclescape() instead of curly brackets -- {} -- since
those only work in most, but not all, cases.

tcllist() now has an optional argument stringify, which specifies the
function used to convert objects to str before being passed to
tclescape(). Defaults to str, but could be a partial(format) call, for
example.