Change ScratchFile to create temporary directories
Backport changes to TemporaryDirectory present in Python 3.12 to support
this.

The main motivation for this change is that it greatly simplifies things
around delete_files: previously, it was pretty manual to keep track of
what files are generated and should be removed. By putting everything in
a temporary directory (within the root scratch directory), the analysis
can just remember the temporary directory -- or even not worry about it
at all, and use the finalization mechanism to schedule deletion to occur
automatically.